feat(photo): enhance photo submission sheet experience
- Add presentationDetents to control sheet size
- Improve photo placeholder with camera icon
